Functions in C
you will learn simple usage of functions in C. Functions are a bunch of statements glued together. A function is provided with zero or more arguments, and it executes the statements on it. Based on the return type, it either returns nothing (void) or something. 

You have to write a function int max_of_four(int a, int b, int c, int d) which reads four arguments and returns the greatest of them.